Included VulkanCube example to test functional end to end Vulkan initialization
