+++ go.mod
@@ -8,6 +8,8 @@ require (
 	github.com/go-gl/gl v0.0.0-20211210172815-726fda9656d6
 	github.com/go-gl/glfw/v3.3/glfw v0.0.0-20240118000515-a250818d05e3
 	github.com/go-gl/mathgl v1.0.0
+	github.com/vulkan-go/asche v0.0.0-20190912134304-d4b318b67e07
+	github.com/vulkan-go/demos v0.0.0-20221129164718-189c453d5553
 	github.com/vulkan-go/vulkan v0.0.0-20221209234627-c0a353ae26c8
 	go.uber.org/zap v1.17.0
 )
@@ -32,7 +34,6 @@ require (
 	github.com/srwiley/rasterx v0.0.0-20220730225603-2ab79fcdd4ef // indirect
 	github.com/stretchr/testify v1.8.4 // indirect
 	github.com/tevino/abool v1.2.0 // indirect
-	github.com/vulkan-go/asche v0.0.0-20190912134304-d4b318b67e07 // indirect
 	github.com/xlab/linmath v0.0.0-20220922225318-40b6290c3b40 // indirect
 	github.com/yuin/goldmark v1.5.5 // indirect
 	go.uber.org/atomic v1.7.0 // indirect
